requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
arrow==1.4.0
attrs==25.4.0
audioop-lts==0.2.2
black==25.12.0
braceexpand==0.1.7
click==8.3.1
discord.py==2.6.4
dnspython==2.8.0
frozenlist==1.8.0
h11==0.16.0
idna==3.11
import_expression==2.2.1.post1
isort==7.0.0
jishaku==2.6.3
lxml==6.0.2
markdown-it-py==4.0.0
mdurl==0.1.2
multidict==6.7.0
mypy_extensions==1.1.0
orjson==3.11.5
packaging==25.0
parsedatetime==2.6
pathspec==0.12.1
platformdirs==4.5.1
propcache==0.4.1
pyaml==25.7.0
Pygments==2.19.2
pymongo==4.15.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-frontmatter==1.1.0
pytokens==0.3.0
PyYAML==6.0.3
RapidFuzz==3.14.3
redis==7.1.0
rich==14.2.0
ruff==0.14.10
six==1.17.0
tabulate==0.9.0
typing_extensions==4.15.0
tzdata==2025.3
uvicorn==0.40.0
uvloop==0.22.1
yarl==1.22.0
